Move Client session handling to separate file
