+++ server/src/testing.rs
@@ -2,6 +2,7 @@ pub mod client;
 mod dispatch;
 pub mod key;
 pub mod server;
+pub mod session;
 
 pub use crate::server::Server;
 pub use crate::testing::client::Client;
